app folder restructured for standardization
