+++ app/dashboard/components/main-nav.tsx
@@ -5,7 +5,7 @@ import { cn } from '@/lib/utils';
 import { signOut } from 'next-auth/react';
 import { usePathname } from 'next/navigation'; // Import the useRouter hook
 
-export function MainNav({
+export default function MainNav({
   className,
   ...props
 }: React.HTMLAttributes<HTMLElement>) {
