+++ app/dashboard/components/user-nav.tsx
@@ -15,7 +15,7 @@ import {
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu"
 
-export function UserNav() {
+export default function UserNav() {
   return (
     <DropdownMenu>
       <DropdownMenuTrigger asChild>
